Provide a keep alive mechanism, split logic, and refactor dispatch
The server may drop the connection if we haven't talked to it in a
while. This commit provides a mechanism for staying alive. The format
is the same as the `:cmds` option and processed in the same way. On
init, the server schedules a future `:keep_alive` message provided the
`:keep_alive_timeout` has been provided. When the server handles the
message, it spawns the keep alive commands and then schedules another
`:keep_alive` message to be sent in `:keep_alive_timeout` milliseconds.

The original startup function had a useful operation that we can reuse
in the `:keep_alive` reaction, so it has been extracted into its own
function `run_cmds/2`.

The dispatch function has been made simpler by just using a
comprehension instead of recursing from the function head.