Leave continuation mode when connection is lost
Specifically when IDLEing, the server is in continuation mode, which
means we are still within the context of a tagged command. Instead, a
continuation command should be sent to the server, i.e. `DONE\r\n`. For
this reason, we change the state when a continuation symbol (`+`)
comes back from the server. When the next command is sent, we send it
to the server without a tag.

Reconnection was failing specifically for the IDLE process. Upon
reconnection, the server was expecting a tagged command (i.e.
`C433 LOGIN user pass\r\n`, but we were erroneously sending an untagged
command i.e. `LOGIN user pass\r\n` (since we were still in
continuation mode after disconnection). That is a BAD command -- the
server thinks `LOGIN` is the command tag, but then the next token is
NOT a command.

This commit resolves that issue by returning the state to `+: false`
when the connection is closed.