Flush commands when disconnected, notify others when socket down
Commands could potentially pile up which would cause many repeat
messages to the same process and waste resources. This commit adds
logic to flush the commands when the connection is lost.

When a command is received and we don't have a socket, dispatch a
message to the original caller notifying them of such.